refactor: 처음 받는 카드의 수 검증 로직 플레이어로 이동
